GOCBC-919: Improve linting abilities
Motivation
----------
We've been using the same linting for quite some time now and IDEs
are beginning to flag up issues that our CI isn't. We've moved gocb
to use golangci-lint so we should do the same for gocbcore.

Changes
-------
Replace all existing linting with golangci-lint which covers the
linters we were using and many more.
Pick a sensible set of linters for now, allowing us to add more in
the future.
Make various code fixes to satisfy the linters.

Reviewed-on: http://review.couchbase.org/c/gocbcore/+/130773
Tested-by: Charles Dixon <chvckd@gmail.com>

+++ memd/conn.go
@@ -139,7 +139,7 @@ func (c *Conn) WritePacket(pkt *Packet) error {
 			return errors.New("cannot specify status in a request packet")
 		}
 
-		binary.BigEndian.PutUint16(buffer[6:], uint16(pkt.Vbucket))
+		binary.BigEndian.PutUint16(buffer[6:], pkt.Vbucket)
 	} else if pkt.Magic == CmdMagicRes {
 		if pkt.Vbucket != 0 {
 			return errors.New("cannot specify vbucket in a response packet")
@@ -242,7 +242,7 @@ func (c *Conn) WritePacket(pkt *Packet) error {
 		serverDurationEnc := EncodeSrvDura16(pkt.ServerDurationFrame.ServerDuration)
 
 		buffer[bodyPos+0] = makeFrameHeader(frameTypeResSrvDuration, 2)
-		binary.BigEndian.PutUint16(buffer[bodyPos+1:], uint16(serverDurationEnc))
+		binary.BigEndian.PutUint16(buffer[bodyPos+1:], serverDurationEnc)
 		bodyPos += 3
 	}
